+++ crates/codexi-cli/src/main.rs
@@ -9,7 +9,6 @@
 
 use anyhow::Result;
 use clap::Parser;
-use log::LevelFilter;
 use std::env;
 
 mod command;
@@ -25,20 +24,8 @@ use codexi::logic::codexi::CodexiSettings;
 use crate::command::Cli;
 use crate::handler::handle_root_command;
 
-fn init_logger(lvl: bool) {
-    // Configuration of the logger
-    let log_level = if lvl { LevelFilter::Debug } else { LevelFilter::Info };
-
-    env_logger::Builder::new()
-        .filter_level(log_level)
-        .format_timestamp_millis()
-        .init();
-}
-
 fn main() -> Result<()> {
     let cli = Cli::parse();
-    // init logger, true = debug
-    init_logger(true);
 
     let settings = CodexiSettings::load_or_create()?;
     let paths = DataPaths::new(&settings.data_dir);